13569: Update dispatchcloud to use arvados.ByteSize.
Arvados-DCO-1.1-Signed-off-by: Tom Clegg <tclegg@veritasgenetics.com>

+++ lib/dispatchcloud/node_size.go
@@ -51,8 +51,8 @@ func ChooseInstanceType(cc *arvados.Cluster, ctr *arvados.Container) (best arvad
 	for _, it := range cc.InstanceTypes {
 		switch {
 		case ok && it.Price > best.Price:
-		case it.Scratch < needScratch:
-		case it.RAM < needRAM:
+		case int64(it.Scratch) < needScratch:
+		case int64(it.RAM) < needRAM:
 		case it.VCPUs < needVCPUs:
 		case it.Preemptible != ctr.SchedulingParameters.Preemptible:
 		case it.Price == best.Price && (it.RAM < best.RAM || it.VCPUs < best.VCPUs):
